+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Sort languages by their translated name
 - Fix the copy to clipboard icon not displayed on Apple systems
 - Close the menu on small screens by Esc
+- Drop compatibility with IIS 5 and with IIS 7.0 before its 2008 FastCGI update
 - Tables overview: speed up, show totals in SQLite, real size in MS SQL, Oracle
 - Table structure: Shorten long checks
 - SQL command: Show column type in result table header